+++ packages/terra-date-time-picker/src/DateTimePicker.jsx
@@ -135,6 +135,11 @@ const propTypes = {
    * If the `variant` prop if set to `12-hour` for one of these supported locales, the variant will be ignored and defaults to `24-hour`.
    */
   timeVariant: PropTypes.oneOf([DateTimeUtils.FORMAT_12_HOUR, DateTimeUtils.FORMAT_24_HOUR]),
+   /**
+   * If invalid error text is used, provide a string containing the IDs for error html element.
+   * ID must be htmlFor prop value with error text.
+   */
+   ariaDescribedBy: PropTypes.string,
 };
 
 const defaultProps = {
@@ -160,6 +165,7 @@ const defaultProps = {
   value: undefined,
   timeVariant: DateTimeUtils.FORMAT_24_HOUR,
   initialTimeZone: DateTimeUtils.getLocalTimeZone(),
+  ariaDescribedBy: '',
 };
 
 class DateTimePicker extends React.Component {
@@ -757,6 +763,7 @@ class DateTimePicker extends React.Component {
       timeInputAttributes,
       value,
       timeVariant,
+      ariaDescribedBy,
       ...customProps
     } = this.props;
 
@@ -809,6 +816,7 @@ class DateTimePicker extends React.Component {
             required={required}
             initialTimeZone={this.initialTimeZone}
             isDefaultDateAcceptable
+            ariaDescribedBy={ariaDescribedBy}
           />
         </div>
         <div className={cx('time-facade')}>
@@ -830,6 +838,7 @@ class DateTimePicker extends React.Component {
             required={required}
             atMaxDate={atMaxDate}
             atMinDate={atMinDate}
+            ariaDescribedBy={ariaDescribedBy}
           />
 
           {this.state.isAmbiguousTime && this.state.dateTime ? this.renderTimeClarification() : null}