+++ packages/terra-time-input/src/TimeInput.jsx
@@ -115,6 +115,11 @@ const propTypes = {
    * If the `variant` prop if set to `12-hour` for one of these supported locales, the variant will be ignored and defaults to `24-hour`.
    */
   variant: PropTypes.oneOf([TimeUtil.FORMAT_12_HOUR, TimeUtil.FORMAT_24_HOUR]),
+  /**
+   * If invalid error text is used, provide a string containing the IDs for error html element.
+   * ID must be htmlFor prop value with error text.
+   */
+  ariaDescribedBy: PropTypes.string,
 };
 
 const defaultProps = {
@@ -136,6 +141,7 @@ const defaultProps = {
   showSeconds: false,
   value: undefined,
   variant: TimeUtil.FORMAT_24_HOUR,
+  ariaDescribedBy: '',
 };
 
 class TimeInput extends React.Component {
@@ -887,6 +893,7 @@ class TimeInput extends React.Component {
       showSeconds,
       value,
       variant,
+      ariaDescribedBy,
       ...customProps
     } = this.props;
 
@@ -1084,6 +1091,7 @@ class TimeInput extends React.Component {
             size="2"
             pattern="\d*"
             description={`${hourDescription()}, ${hotKeyDescription}`}
+            ariaDescribedBy={ariaDescribedBy}
           />
           <TimeSpacer className={cx('time-spacer')} />
           <AccessibleInput
@@ -1104,6 +1112,7 @@ class TimeInput extends React.Component {
             pattern="\d*"
             // description: Like the hour descriptions, but for the minute input.
             description={`${minuteDescription}, ${hotKeyDescription}`}
+            ariaDescribedBy={ariaDescribedBy}
           />
           {showSeconds && (
           <React.Fragment>
@@ -1126,6 +1135,7 @@ class TimeInput extends React.Component {
               pattern="\d*"
               // description: Like the hour descriptions, but for the second input.
               description={`${secondDescription}, ${hotKeyDescription}`}
+              ariaDescribedBy={ariaDescribedBy}
             />
           </React.Fragment>
           )}