+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Added
+  * Added ariaDescribedBy prop in `terra-date-time-picker` example for invalid error message.
+
+* Added
   * Added shouldTrapFocus to be set to 'true'.
 * Changed
   * Updated the `terra-date-time-picker` example for field label.